# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Документация StudX API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080/api",
            "description": "Локальный сервер"
        }
    ],
    "paths": {
        "/courses": {
            "get": {
                "tags": [
                    "courses"
                ],
                "summary": "Получение списка всех курсов",
                "operationId": "22e2add4714520e5d7420442140cffcb",
                "responses": {
                    "200": {
                        "description": "Успешное получение списка всех курсов",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/courses"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Ошибка сервера"
                                        },
                                        "details": {
                                            "type": "string",
                                            "example": "error details"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/courses/{id}": {
            "get": {
                "tags": [
                    "courses"
                ],
                "summary": "Получение информации о конкретном курсе",
                "operationId": "0b9b3dc0b0328a34f28f8b6c91152a68",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Идентификатор курса",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешное получение информации о курсе",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/courses"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Курс не найден",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Курс не найден"
                                        },
                                        "details": {
                                            "type": "string",
                                            "example": "error details"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Ошибка сервера"
                                        },
                                        "details": {
                                            "type": "string",
                                            "example": "error details"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "courses": {
                "title": "Курс",
                "description": "Схема курса",
                "required": [
                    "id",
                    "code",
                    "name"
                ],
                "properties": {
                    "id": {
                        "description": "Идентификатор курса",
                        "type": "integer",
                        "example": 1
                    },
                    "code": {
                        "description": "Код курса",
                        "type": "string",
                        "example": "WEB102"
                    },
                    "name": {
                        "description": "Название курса",
                        "type": "string",
                        "example": "Responsive Web Design"
                    },
                    "subject_name": {
                        "description": "Название предмета",
                        "type": "string",
                        "example": "UI/UX Design",
                        "nullable": true
                    },
                    "created_at": {
                        "description": "Дата и время создания курса",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-12-19T12:34:56Z"
                    },
                    "updated_at": {
                        "description": "Дата и время последнего обновления курса",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-12-19T12:34:56Z"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}